Change CORS Proxy, Fix Response Resolve to Text.
Proxy change was necessary due to improper fetch of data, because the querying URL was omitted.
Response was fixed to text.
Render a Text to the Content Field Element.

Note: Because author/owner of  https://github.com/Rob--W/cors-anywhere/issues/301
Has limited the usage of  PROXY CORS, a self-hosted Proxy or use of another Proxy must me implemented!